+++ nobodywho/flutter/nobodywho/lib/src/rust/frb_generated.dart
@@ -409,6 +409,9 @@ abstract class NobodyWhoApi extends BaseApi {
     double? speed = null,
     int? steps = null,
     double? silenceDuration = null,
+    String? precision = null,
+    double? temperature = null,
+    String? huggingfaceToken = null,
     bool useGpu = true,
   });
 
@@ -3282,6 +3285,9 @@ class NobodyWhoApiImpl extends NobodyWhoApiImplPlatform
     double? speed = null,
     int? steps = null,
     double? silenceDuration = null,
+    String? precision = null,
+    double? temperature = null,
+    String? huggingfaceToken = null,
     bool useGpu = true,
   }) {
     return handler.executeNormal(
@@ -3295,6 +3301,9 @@ class NobodyWhoApiImpl extends NobodyWhoApiImplPlatform
           sse_encode_opt_box_autoadd_f_32(speed, serializer);
           sse_encode_opt_box_autoadd_u_32(steps, serializer);
           sse_encode_opt_box_autoadd_f_32(silenceDuration, serializer);
+          sse_encode_opt_String(precision, serializer);
+          sse_encode_opt_box_autoadd_f_32(temperature, serializer);
+          sse_encode_opt_String(huggingfaceToken, serializer);
           sse_encode_bool(useGpu, serializer);
           pdeCallFfi(
             generalizedFrbRustBinding,
@@ -3317,6 +3326,9 @@ class NobodyWhoApiImpl extends NobodyWhoApiImplPlatform
           speed,
           steps,
           silenceDuration,
+          precision,
+          temperature,
+          huggingfaceToken,
           useGpu,
         ],
         apiImpl: this,
@@ -3334,6 +3346,9 @@ class NobodyWhoApiImpl extends NobodyWhoApiImplPlatform
       "speed",
       "steps",
       "silenceDuration",
+      "precision",
+      "temperature",
+      "huggingfaceToken",
       "useGpu",
     ],
   );